#3: Extract only common constants from FE and BE
