+++ index.js
@@ -1,4 +1,4 @@
 module.exports = {
   utils: require('./utils'),
-  client: require('./client')
+  constants: require('./constants')
 };
